requirements.txt
pytesseract==0.3.13
pillow
pandas
pdf2image